Added ClassFinishInstanceStep
Allows class object to finalise the instance that has just been created from it.
-This does the freezing and allows dynamic to simply overload and undo the freeze.
