+++ .gitignore
@@ -74,3 +74,4 @@ InitTestScene*.unity
 # End of https://www.toptal.com/developers/gitignore/api/unity
 CodeCoverage/
 .vsconfig
+ProjectSettings/Packages/com.unity.testtools.codecoverage/Settings.json
