+++ Assets/ulox/Tests/ClassInherTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9e661c4238f0c9f429107365b5c1d5b0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
