+++ Assets/ulox/Tests/ClassStaticTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d14ce673bf38fbb47b7354c4160bea2d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
