+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2021.1.21f1
-m_EditorVersionWithRevision: 2021.1.21f1 (f2d5d3c59f8c)
+m_EditorVersion: 2021.1.23f1
+m_EditorVersionWithRevision: 2021.1.23f1 (b50729e604a9)
